# apps/api/src/agentprobe_api/runs.py
"""Runs (SPEC.md §8, ADR 0017): start a suite run, read it, cancel it, and stream its
progress over Server-Sent Events, with the stream-token fallback of ADR 0009 §5.

Starting a run stores a config snapshot (suite YAML, agent config, secret id) and hands
the run to the queue backend; how it executes is core's `agentprobe_core.runner`.
"""

import json
import uuid
from collections.abc import AsyncIterator
from contextlib import AsyncExitStack
from datetime import UTC, datetime
from typing import Any, Literal

from fastapi import APIRouter, Query, Request, Response
from fastapi.responses import HTMLResponse, JSONResponse, StreamingResponse
from pydantic import BaseModel, ConfigDict, Field
from sqlalchemy import func, select, update
from sqlalchemy.ext.asyncio import AsyncSession

from agentprobe_api import limits, runstore
from agentprobe_api.auth import (
    STREAM_TOKEN_TTL,
    AppSettings,
    CurrentPrincipal,
    CurrentUser,
    Db,
    Principal,
    get_principal,
    mint_stream_token,
    stream_principal,
)
from agentprobe_api.errors import ApiError
from agentprobe_api.htmlexport import render_html
from agentprobe_api.models import Agent, Project, Run, RunCaseSummary, Suite
from agentprobe_api.progress import Subscription
from agentprobe_api.projects import owned_project
from agentprobe_api.runstore import LIVE, TERMINAL
from agentprobe_api.suites import owned_suite
from agentprobe_core.runner import plan_attempts
from agentprobe_core.suite import SuiteParseError, parse_suite_yaml

router = APIRouter(tags=["runs"])
HEARTBEAT_S = 15.0  # an SSE comment this often keeps proxies from closing an idle stream
EXPORT_HEADERS = {
    "Content-Security-Policy": (
        "default-src 'none'; style-src 'unsafe-inline'; base-uri 'none'; form-action 'none'"
    ),
    "X-Content-Type-Options": "nosniff",
}


class RunIn(BaseModel):
    model_config = ConfigDict(extra="forbid")
    runs_per_case: int | None = Field(default=None, ge=1, le=20)  # default: the suite's
    model: str | None = Field(default=None, max_length=200)  # a label for the agent's model
    mock: bool = True  # mock LLM for judges; false uses the server's LLM_PROVIDER
    git_sha: str | None = Field(default=None, max_length=64)
    branch: str | None = Field(default=None, max_length=200)
    pr_number: int | None = Field(default=None, ge=1)


class RunOut(BaseModel):
    model_config = ConfigDict(from_attributes=True)
    id: uuid.UUID
    suite_id: uuid.UUID
    suite_version: int
    agent_id: uuid.UUID | None  # None: an unregistered agent, named by agent_name (ADR 0020)
    agent_name: str | None
    status: str
    model: str | None
    git_sha: str | None
    branch: str | None
    pr_number: int | None
    runs_per_case: int
    mock_mode: bool
    attempts_total: int
    attempts_done: int
    pass_rate: float | None
    ci_lower: float | None
    ci_upper: float | None
    total_tokens: int | None
    total_cost: float | None
    judge_cost_usd: float | None
    error: str | None
    created_at: datetime
    started_at: datetime | None
    finished_at: datetime | None


class RunDetailOut(RunOut):
    """A run for its own page: the suite's name, the agent's name as the run recorded it, and
    whether a share link is live (the link's token is only ever returned when it's created).
    """

    suite_name: str
    agent: str
    shared: bool
    share_expires_at: datetime | None


class StreamTokenOut(BaseModel):
    token: str
    expires_in: int


class RunListOut(BaseModel):
    """A run as shown in a project's run history (overview trends, latest-runs table).
    `agent_name` is the agent's current name for a registered agent, or the name an
    unregistered/CI-pushed run recorded for itself (ADR 0020) — whichever applies.
    `mean_latency_ms` isn't a stored column: it's the attempts-weighted average of
    `run_case_summaries.mean_latency_ms` across the run's cases, computed at read time.
    """

    id: uuid.UUID
    suite_id: uuid.UUID
    suite_name: str
    agent_id: uuid.UUID | None
    agent_name: str | None
    status: str
    branch: str | None
    mock_mode: bool
    pass_rate: float | None
    ci_lower: float | None
    ci_upper: float | None
    total_cost: float | None
    judge_cost_usd: float | None
    total_tokens: int | None
    mean_latency_ms: float | None
    created_at: datetime
    finished_at: datetime | None


async def owned_run(db: AsyncSession, principal: Principal, run_id: uuid.UUID) -> Run:
    """Not-owned and nonexistent are both 404 (mirrors `owned_suite`)."""
    row = (
        await db.execute(
            select(Run, Suite.project_id)
            .join(Suite, Suite.id == Run.suite_id)
            .join(Project, Project.id == Suite.project_id)
            .where(Run.id == run_id, Project.user_id == principal.user_id)
            .execution_options(populate_existing=True)
        )
    ).first()
    if row is None or (principal.project_id is not None and row[1] != principal.project_id):
        raise ApiError(404, "Run not found")
    run: Run = row[0]
    return run


@router.post("/suites/{suite_id}/runs", status_code=202)
async def start_run(
    suite_id: uuid.UUID,
    body: RunIn,
    principal: CurrentPrincipal,
    db: Db,
    request: Request,
    settings: AppSettings,
) -> RunOut:
    suite = await owned_suite(db, principal, suite_id)
    try:
        parsed = parse_suite_yaml(suite.yaml_source)
    except SuiteParseError as exc:  # stored YAML was valid when uploaded; limits may change
        raise ApiError(
            422,
            "Suite validation failed",
            details=[issue.model_dump(mode="json") for issue in exc.issues],
        ) from exc
    agent = await db.scalar(
        select(Agent).where(Agent.project_id == suite.project_id, Agent.name == parsed.agent)
    )
    if agent is None:
        raise ApiError(422, f"The suite's agent {parsed.agent!r} isn't in this project")
    if agent.adapter_type != "http":
        raise ApiError(422, f"{agent.adapter_type} agents can't run on the server yet")
    await limits.check_runs(db, settings, principal.user_id)
    await limits.check_live_budget(db, settings, mock=body.mock)
    runs_per_case = body.runs_per_case or parsed.runs_per_case
    attempts = plan_attempts(parsed, runs_per_case)
    run = Run(
        suite_id=suite.id,
        suite_version=suite.version,
        agent_id=agent.id,
        status="queued",
        model=body.model,
        git_sha=body.git_sha,
        branch=body.branch,
        pr_number=body.pr_number,
        runs_per_case=runs_per_case,
        mock_mode=body.mock,
        config_snapshot=runstore.make_snapshot(
            suite, agent, runs_per_case=runs_per_case, mock=body.mock
        ),
        attempts_total=len(attempts),
    )
    db.add(run)
    await db.flush()
    out = RunOut.model_validate(run)
    await db.commit()  # the queue runs it in another session: it must be visible first
    await request.app.state.queue.enqueue(run.id)
    return out


@router.get("/projects/{project_id}/runs")
async def list_project_runs(
    project_id: uuid.UUID,
    principal: CurrentPrincipal,
    db: Db,
    suite_id: uuid.UUID | None = None,
    limit: int = Query(default=50, ge=1, le=200),
    offset: int = Query(default=0, ge=0),
) -> list[RunListOut]:
    project = await owned_project(db, principal, project_id)
    latency = (
        select(
            func.sum(RunCaseSummary.mean_latency_ms * RunCaseSummary.attempts)
            / func.nullif(func.sum(RunCaseSummary.attempts), 0)
        )
        .where(RunCaseSummary.run_id == Run.id, RunCaseSummary.mean_latency_ms.isnot(None))
        .correlate(Run)
        .scalar_subquery()
    )
    stmt = (
        select(Run, Suite.name, Agent.name, latency)
        .join(Suite, Suite.id == Run.suite_id)
        .outerjoin(Agent, Agent.id == Run.agent_id)
        .where(Suite.project_id == project.id)
    )
    if suite_id is not None:
        stmt = stmt.where(Run.suite_id == suite_id)
    stmt = stmt.order_by(Run.created_at.desc()).limit(limit).offset(offset)
    rows = await db.execute(stmt)
    return [
        RunListOut(
            id=run.id,
            suite_id=run.suite_id,
            suite_name=suite_name,
            agent_id=run.agent_id,
            agent_name=agent_name if agent_name is not None else run.agent_name,
            status=run.status,
            branch=run.branch,
            mock_mode=run.mock_mode,
            pass_rate=run.pass_rate,
            ci_lower=run.ci_lower,
            ci_upper=run.ci_upper,
            total_cost=float(run.total_cost) if run.total_cost is not None else None,
            judge_cost_usd=float(run.judge_cost_usd) if run.judge_cost_usd is not None else None,
            total_tokens=run.total_tokens,
            mean_latency_ms=float(mean_latency_ms) if mean_latency_ms is not None else None,
            created_at=run.created_at,
            finished_at=run.finished_at,
        )
        for run, suite_name, agent_name, mean_latency_ms in rows.tuples()
    ]


@router.get("/runs/{run_id}")
async def get_run(run_id: uuid.UUID, principal: CurrentPrincipal, db: Db) -> RunDetailOut:
    run = await owned_run(db, principal, run_id)
    suite_name = (await db.execute(select(Suite.name).where(Suite.id == run.suite_id))).scalar_one()
    expired = run.share_expires_at is not None and run.share_expires_at <= datetime.now(UTC)
    return RunDetailOut(
        **RunOut.model_validate(run).model_dump(),
        suite_name=suite_name,
        agent=str(run.config_snapshot["agent"]["name"]),
        shared=run.share_token_hash is not None and not expired,
        share_expires_at=run.share_expires_at,
    )


@router.post("/runs/{run_id}/cancel")
async def cancel_run(
    run_id: uuid.UUID, principal: CurrentPrincipal, db: Db, request: Request
) -> RunOut:
    run = await owned_run(db, principal, run_id)
    changed = await db.scalar(
        update(Run)
        .where(Run.id == run.id, Run.status.in_(LIVE))
        .values(status="cancelled", finished_at=func.now())
        .returning(Run.id)
    )
    if changed is None:
        raise ApiError(409, f"The run has already ended ({run.status})")
    await db.commit()
    await db.refresh(run)
    await request.app.state.queue.cancel(run.id)
    await request.app.state.bus.publish(run.id, runstore.run_event(run))
    return RunOut.model_validate(run)


@router.post("/runs/{run_id}/stream-token")
async def stream_token(
    run_id: uuid.UUID, principal: CurrentUser, db: Db, settings: AppSettings
) -> StreamTokenOut:
    run = await owned_run(db, principal, run_id)
    return StreamTokenOut(
        token=mint_stream_token(settings, principal.user_id, run.id),
        expires_in=int(STREAM_TOKEN_TTL.total_seconds()),
    )


@router.get("/runs/{run_id}/stream")
async def stream_run(
    run_id: uuid.UUID,
    request: Request,
    db: Db,
    settings: AppSettings,
    token: str | None = None,
) -> StreamingResponse:
    """`snapshot` (the run's state now), then `attempt` and `status` events until the run
    ends. Authenticated like any request, or with `?token=` from `POST
    /runs/{id}/stream-token` where the browser can't send the session cookie.
    """
    principal = (
        stream_principal(settings, token, run_id)
        if token is not None
        else await get_principal(request, db, settings)
    )
    await owned_run(db, principal, run_id)
    stack = AsyncExitStack()
    # Subscribe before reading the snapshot, so no event falls between the two.
    subscription = await stack.enter_async_context(request.app.state.bus.subscribe(run_id))
    try:
        run = await owned_run(db, principal, run_id)
        snapshot = runstore.run_event(run, "snapshot")
    except BaseException:
        await stack.aclose()
        raise
    headers = {"Cache-Control": "no-cache", "X-Accel-Buffering": "no"}
    if token is not None:
        # The browser connects here directly, cross-origin, because the web app's /api rewrite
        # gzips and so buffers an event stream (ADR 0031). Only the web origin may read it; no
        # credentials are involved, the token is the whole authorization.
        headers |= {"Access-Control-Allow-Origin": settings.web_origin, "Vary": "Origin"}
    return StreamingResponse(
        _events(snapshot, subscription, stack), media_type="text/event-stream", headers=headers
    )


def _sse(event: dict[str, Any]) -> str:
    return f"event: {event['type']}\ndata: {json.dumps(event, separators=(',', ':'))}\n\n"


async def _events(
    snapshot: dict[str, Any], subscription: Subscription, stack: AsyncExitStack
) -> AsyncIterator[str]:
    try:
        yield _sse(snapshot)
        if snapshot["status"] in TERMINAL:
            return
        while True:
            event = await subscription.next(HEARTBEAT_S)
            if event is None:
                yield ": keep-alive\n\n"
                continue
            yield _sse(event)
            if event["type"] == "status" and event["status"] in TERMINAL:
                return
    finally:
        await stack.aclose()


@router.get("/runs/{run_id}/export")
async def export_run(
    run_id: uuid.UUID,
    principal: CurrentPrincipal,
    db: Db,
    format: Literal["json", "html"] = "json",
) -> Response:
    """A portable record of the run, recomputed from the persisted attempts via core's own
    `finalize_run` (ADR 0018): the strongest guarantee it agrees with core's statistics.
    `format=html` is one self-contained, fully escaped file (no external assets).
    """
    run = await owned_run(db, principal, run_id)
    summary = await runstore.read_summary(db, run)
    if format == "html":
        # Served on the app's origin (the /api rewrite): even a missed escape can't run a
        # script, load anything or submit anywhere.
        return HTMLResponse(render_html(run, summary), headers=EXPORT_HEADERS)
    return JSONResponse(
        {
            "run": RunOut.model_validate(run).model_dump(mode="json"),
            "summary": summary.model_dump(mode="json"),
        }
    )